Implemented an assertion library with various failure handling options
This commit introduces the `AssertionError` class, which represents an assertion
failure and stores information about the failed expression, file, and line number.
It also introduces the following functions:
- `__Assert`, which prints an error message and aborts the program on failure
- `__Assert_w_Err_Class`, which throws an `AssertionError` object on failure
- `__Assert_Long_Jmp`, which stores an `AssertionError` object in a global variable
  and jumps back to a `setjmp` call in the calling function on failure

In addition, this commit introduces the following macros:
- `ASSERTIFY_ASSERT_ABORT`, which calls `__Assert` on failure
- `ASSERTIFY_ASSERT_EXCEPTION`, which throws an `AssertionError` object on failure
  using `__Assert_w_Err_Class` or `__Assert_Long_Jmp` (depending on the value of
  `ASSERTIFY_LONG_JMP_ENDABLED`)

These functions and macros allow the user to choose how to handle an assertion
failure, depending on their needs and preferences.